Updated the logic to remove explicit conversion

+++ Sum of n numbers/SumOfNNumbers.java
@@ -4,6 +4,6 @@ class SumOfNNumber{
     public static void main(String args[]){
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        System.out.println((int)(n*(n+1)/2));
+        System.out.println(((n*(n+1))/2));
     }
 }
